requirements.txt
fastapi
uvicorn[standard]
pymongo
redis
requests
python-dotenv

python-multipart
torch
transformers
scikit-learn
python-docx
openpyxl
pytesseract
Pillow
pdfminer.six
celery[redis]
torch
torchvision
pytest
httpx
SQLAlchemy